Backports to 1.7.1:
add the -pem option in extract-signature mode

This changes the extracted output to plaintext PEM format.

add the attach-signature command

Combine a previously extracted signature with an unsigned file and output a
valid, signed result. Accepts binary or pem pkcs7 inputs.

This is helpful for a deterministic build process, where the signer is only
required to provide a detached signature and users or other builders can use
it to create an otherwise deterministic binary.